# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: |
    ClinicFlow REST contract. Sessions resolve a database role. The HttpOnly, Secure,
    SameSite=Lax cookie contains an opaque native token by default for continuity,
    or an app-issued JWT when AUTH_SESSION_MODE=jwt is explicitly configured.
    Both modes are database-revocable with a fixed 12-hour lifetime, with no refresh
    endpoint. Changing modes invalidates old cookies. JWT_SIGNING_KEY must be a
    dedicated high-entropy secret of at least 32 UTF-8 bytes, configured before
    enabling JWT mode. Session issuance may return 503 AUTH_SESSION_UNCONFIGURED.
    Newly created passwords require at least 8 characters, ASCII letters and numbers,
    and at most 1024 UTF-8 bytes (enforced by the server).
    Super admins access all; clinic admins and receptionists are assignment-scoped;
    doctors access their own resources and relevant patients; patients access self only.
    All filters narrow authorized scope, never expand it. DELETE deactivates records
    with history. Lists contain real persisted data only. Queue clients poll every 30 seconds.
    Invalid input returns 400, unauthenticated 401, forbidden 403, missing 404,
    state/capacity conflicts 409, rate limiting 429, unavailable integrations 503.
servers:
  - url: /api
security:
  - NativeSession: []
paths:
...
  /public/qr/{reference}:
    get:
      operationId: resolveQr
      security: []
      parameters: [{ name: reference, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Active booking context without personal data, content: { application/json: { schema: { $ref: "#/components/schemas/QrContext" } } } }
        "404": { $ref: "#/components/responses/Error" }
...
  /appointments:
    get:
      operationId: listAppointments
      parameters: [{ $ref: "#/components/parameters/SessionId" }, { $ref: "#/components/parameters/StartTime" }, { $ref: "#/components/parameters/StatusGroup" }, { $ref: "#/components/parameters/Search" }, { $ref: "#/components/parameters/ClinicId" }, { $ref: "#/components/parameters/BranchId" }, { $ref: "#/components/parameters/DoctorId" }, { name: patientId, in: query, schema: { type: string } }, { $ref: "#/components/parameters/Date" }, { $ref: "#/components/parameters/From" }, { $ref: "#/components/parameters/To" }, { name: status, in: query, schema: { $ref: "#/components/schemas/AppointmentStatus" } }, { name: source, in: query, schema: { $ref: "#/components/schemas/BookingSource" } }, { $ref: "#/components/parameters/Page" }, { $ref: "#/components/parameters/PageSize" }, { $ref: "#/components/parameters/Sort" }]
      responses:
        "200": { description: Scoped history and upcoming visits, content: { application/json: { schema: { $ref: "#/components/schemas/AppointmentList" } } } }
    post:
      operationId: createAppointment
      description: Atomically checks assignment, computed availability, OTP policy and capacity, allocates reference/token. Patients can book only self; staff may create walk-in and phone bookings. Duplicate requestId returns original booking.
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/AppointmentInput" } } } }
      responses:
        "201": { description: Booking confirmation, content: { application/json: { schema: { $ref: "#/components/schemas/Appointment" } } } }
        "409": { $ref: "#/components/responses/Error" }
  /appointments/{id}:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      operationId: getAppointment
      responses:
        "200": { description: Confirmation and lifecycle, content: { application/json: { schema: { $ref: "#/components/schemas/Appointment" } } } }
...
  /appointments/{id}/reschedule:
    post:
      operationId: rescheduleAppointment
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/AppointmentReschedule" } } } }
      responses:
        "200": { description: Rescheduled atomically, content: { application/json: { schema: { $ref: "#/components/schemas/Appointment" } } } }
  /appointments/{id}/actions:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    post:
      operationId: transitionAppointment
      description: |
        New bookings immediately enter waiting with a stable token.
        checkIn called (or first pending with no current patient)→inConsultation,
        stamping checkedInAt and consultationStartedAt.
        complete inConsultation→completed and atomically calls the next pending reservation.
        Legacy booked/checkedIn reservations remain usable without migration.
        Compatibility enqueue normalizes legacy pending to waiting; call pending→called; start called→inConsultation.
        noShow pending/called→noShow requires reason; requeue noShow→waiting requires reason, position and fresh versions.
        cancel booked/checkedIn/waiting/called→cancelled only before checkedInAt and within cutoff.
        Queue actions are appointment-date-only, serialized per session, and never automatically skip patients.
        Patients may cancel only self. All other actions require scoped staff.
        Concurrent state conflicts return 409. Every transition is audited.
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/AppointmentAction" } } } }
      responses:
        "200": { description: Updated lifecycle, content: { application/json: { schema: { $ref: "#/components/schemas/Appointment" } } } }
        "409": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    NativeSession:
      type: apiKey
      in: cookie
      name: digiq_session
      description: Opaque native token (default) or explicitly enabled app-issued JWT. Both are database-revocable, fixed 12-hour HttpOnly/Secure/SameSite=Lax cookies. No refresh endpoint.
  responses:
    Error:
      description: Request failed
      content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
  parameters:
    SessionId: { name: sessionId, in: query, schema: { type: string } }
    StartTime: { name: startTime, in: query, schema: { type: string } }
...
    Id: { name: id, in: path, required: true, schema: { type: string } }
    Search: { name: search, in: query, schema: { type: string } }
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    Sort: { name: sort, in: query, description: Allowlisted field with optional minus prefix for descending order, schema: { type: string } }
...
    StatusGroup: { name: statusGroup, in: query, description: Server-side group filter applied before pagination; intersects with status when both supplied, schema: { type: string, enum: [active, waiting, absent, completed, cancelled, all] } }
    ClinicId: { name: clinicId, in: query, schema: { type: string } }
    BranchId: { name: branchId, in: query, schema: { type: string } }
    DoctorId: { name: doctorId, in: query, schema: { type: string } }
    Date: { name: date, in: query, schema: { type: string, format: date } }
    From: { name: from, in: query, schema: { type: string, format: date } }
    To: { name: to, in: query, schema: { type: string, format: date } }
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }
...
    AppointmentStatus: { type: string, enum: [booked, checkedIn, waiting, called, inConsultation, completed, noShow, cancelled] }
    BookingSource: { type: string, enum: [online, walkIn, phone, qr] }
    PageMeta:
      type: object
      required: [total, page, pageSize]
      properties:
        total: { type: integer, minimum: 0 }
        page: { type: integer }
        pageSize: { type: integer }
        totalPages: { type: integer }
...
    ClinicDateFormat:
      type: string
      enum: ["DD MMM YYYY", "DD/MM/YYYY", "MM/DD/YYYY", "YYYY-MM-DD"]
      description: Parent Clinic Group display preference; locations inherit it. Default DD MMM YYYY.
    ClinicTimeFormat:
      type: string
      enum: ["12h", "24h"]
      description: Parent Clinic Group display preference; locations inherit it. Default 12h.
...
    AppointmentInput:
      type: object
      required: [patientId, doctorId, clinicId, branchId, date, source]
      properties:
        sessionId: { type: string }
        startTime: { type: string }
        patientId: { type: string }
        doctorId: { type: string }
        clinicId: { type: string }
        branchId: { type: string }
        date: { type: string, format: date }
        source: { $ref: "#/components/schemas/BookingSource" }
        appointmentTypeId: { type: string }
        consultationTypeId: { type: string }
        qrReference: { type: string }
        requestId: { type: string, description: Client-generated idempotency key }
        termsAccepted: { type: boolean }
        notes: { type: string, maxLength: 1000 }
    ConfirmationEmailOutcome:
      type: string
      enum: [provider_accepted, unavailable, disabled, no_recipient, not_attempted]
      description: Booking confirmation email attempt outcome. Provider acceptance is not proof of inbox delivery.
    Appointment:
      allOf:
        - { $ref: "#/components/schemas/AppointmentInput" }
        - type: object
          required: [id, reference, token, status, createdAt, patientName, doctorName, clinicName, branchName, allowedActions]
          properties:
            confirmationEmail: { $ref: "#/components/schemas/ConfirmationEmailOutcome" }
            dateFormat: { $ref: "#/components/schemas/ClinicDateFormat" }
            timeFormat: { $ref: "#/components/schemas/ClinicTimeFormat" }
            id: { type: string }
            reference: { type: string }
            token: { type: string }
            tokenNumber: { type: integer }
            queueRank: { type: number }
            revision: { type: integer }
            expectedDurationMinutes: { type: [integer, "null"] }
            status: { $ref: "#/components/schemas/AppointmentStatus" }
            patientName: { type: string }
            patientCode: { type: string }
            doctorName: { type: string }
            clinicName: { type: string }
            branchName: { type: string }
            branchAddress: { type: [string, "null"], description: "Booking branch address snapshot; for legacy bookings the current branch address is resolved." }
            timezone: { type: [string, "null"], description: "Time zone for the consulting session range." }
            startTime: { type: string }
            endTime: { type: string }
            createdAt: { type: string, format: date-time }
            checkedInAt: { type: [string, "null"], format: date-time }
            calledAt: { type: [string, "null"], format: date-time }
            consultationStartedAt: { type: [string, "null"], format: date-time }
            completedAt: { type: [string, "null"], format: date-time }
            allowedActions: { type: array, items: { $ref: "#/components/schemas/AppointmentActionType" } }
            history: { type: array, items: { $ref: "#/components/schemas/StatusEvent" } }
...
    AppointmentActionType: { type: string, enum: [checkIn, enqueue, call, start, complete, noShow, requeue, cancel] }
    AppointmentAction:
      type: object
      required: [action]
      properties:
        action: { $ref: "#/components/schemas/AppointmentActionType" }
        reason: { type: string }
        cancellationReasonId: { type: string }
        expectedStatus: { $ref: "#/components/schemas/AppointmentStatus" }
        expectedRevision: { type: integer, minimum: 0 }
        expectedQueueVersion: { type: string }
        position: { type: integer, minimum: 1, description: Required for requeue. One-based position among pending reservations; current consultation is never displaced. }
    StatusEvent:
      type: object
      required: [status, occurredAt]
      properties:
        status: { $ref: "#/components/schemas/AppointmentStatus" }
        occurredAt: { type: string, format: date-time }
        reason: { type: string }
        action: { type: string }
        position: { type: integer }
        from: { $ref: "#/components/schemas/AppointmentSessionHistory" }
        to: { $ref: "#/components/schemas/AppointmentSessionHistory" }
    AppointmentSessionHistory:
      type: object
      properties:
        doctorId: { type: string }
        branchId: { type: string }
        date: { type: string, format: date }
        token: { type: string }
        tokenNumber: { type: integer }
...
    StatusCounts:
      type: object
      properties:
        active: { type: integer }
        waiting: { type: integer }
        absent: { type: integer }
        completed: { type: integer }
        cancelled: { type: integer }
        all: { type: integer }
...
    AppointmentReschedule:
      type: object
      required: [doctorId, branchId, date, expectedRevision]
      properties:
        sessionId: { type: string }
        startTime: { type: string }
        doctorId: { type: string }
        branchId: { type: string }
        date: { type: string, format: date }
        expectedRevision: { type: integer, minimum: 0 }
        reason: { type: string }
...
    QrContext:
      type: object
      required: [reference, clinicId, clinicName]
      properties:
        dateFormat: { $ref: "#/components/schemas/ClinicDateFormat" }
        timeFormat: { $ref: "#/components/schemas/ClinicTimeFormat" }
        reference: { type: string }
        clinicId: { type: string }
        clinicName: { type: string }
        clinicAddress: { type: [string, "null"] }
        branchAddress: { type: [string, "null"] }
        branchCity: { type: [string, "null"] }
        branchTimezone: { type: [string, "null"] }
        branchId: { type: [string, "null"] }
        branchName: { type: [string, "null"] }
        doctorId: { type: [string, "null"] }
        doctorName: { type: [string, "null"] }
...
    AppointmentList:
      allOf:
        - { $ref: "#/components/schemas/PageMeta" }
        - { type: object, required: [items], properties: { statusCounts: { $ref: "#/components/schemas/StatusCounts" }, items: { type: array, items: { $ref: "#/components/schemas/Appointment" } } } }
...