# lib/api-zod/src/generated/types/clinicDateFormat.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * ClinicFlow REST contract. Sessions resolve a database role. The HttpOnly, Secure,
 * SameSite=Lax cookie contains an opaque native token by default for continuity,
 * or an app-issued JWT when AUTH_SESSION_MODE=jwt is explicitly configured.
 * Both modes are database-revocable with a fixed 12-hour lifetime, with no refresh
 * endpoint. Changing modes invalidates old cookies. JWT_SIGNING_KEY must be a
 * dedicated high-entropy secret of at least 32 UTF-8 bytes, configured before
 * enabling JWT mode. Session issuance may return 503 AUTH_SESSION_UNCONFIGURED.
 * Newly created passwords require at least 8 characters, ASCII letters and numbers,
 * and at most 1024 UTF-8 bytes (enforced by the server).
 * Super admins access all; clinic admins and receptionists are assignment-scoped;
 * doctors access their own resources and relevant patients; patients access self only.
 * All filters narrow authorized scope, never expand it. DELETE deactivates records
 * with history. Lists contain real persisted data only. Queue clients poll every 30 seconds.
 * Invalid input returns 400, unauthenticated 401, forbidden 403, missing 404,
 * state/capacity conflicts 409, rate limiting 429, unavailable integrations 503.
 *
 * OpenAPI spec version: 1.0.0
 */

/**
 * Parent Clinic Group display preference; locations inherit it. Default DD MMM YYYY.
 */
export type ClinicDateFormat = typeof ClinicDateFormat[keyof typeof ClinicDateFormat];


export const ClinicDateFormat = {
  DD_MMM_YYYY: 'DD MMM YYYY',
  'DD/MM/YYYY': 'DD/MM/YYYY',
  'MM/DD/YYYY': 'MM/DD/YYYY',
  'YYYY-MM-DD': 'YYYY-MM-DD',
} as const;